MISKMATCH_BACKEND_FINAL/miskmatch/requirements.txt
# ─────────────────────────────────────────────
# MiskMatch — FastAPI Backend Dependencies
# "Sealed with musk." — ختامه مسك — Quran 83:26
# ─────────────────────────────────────────────

# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
psycopg2-binary==2.9.9

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# Validation
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
phonenumbers==8.13.45

# Cache & Queue
redis==5.1.1
celery==5.4.0
flower==2.0.1

# File Storage & Media
boto3==1.35.27
Pillow==10.4.0

# AI / ML
openai==1.47.0
anthropic==0.34.2
sentence-transformers==3.1.1

# Real-time
websockets==13.1

# HTTP Client
httpx==0.27.2
aiohttp==3.10.8

# Payments
stripe==10.12.0

# Communications
twilio==9.3.0
sendgrid==6.11.0

# Biometric Verification
# onfido-python==5.0.0  # uncomment when adding verification

# Utils
python-slugify==8.0.4
babel==2.16.0  # Arabic/RTL locale support
pytz==2024.2
Faker==30.3.0  # for seeding

# Monitoring
sentry-sdk[fastapi]==2.14.0
prometheus-fastapi-instrumentator==7.0.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
factory-boy==3.3.1

# Dev
black==24.10.0
ruff==0.6.9
mypy==1.11.2
pre-commit==3.8.0
